Removed daycount slots from Term
Term is straight and dumb, at least for now.
The main issue on this decision relates the dificulty in pick up a good approach to compare Terms.

If the terms have the same daycount it is straightforward, but if they don't, what should be done? Raise an error, a warning, return NA.

For this reason, the Compare method has also been removed.
If user needs to compare two terms he must be sure of what he wants.