Added propagate slot to Interpolation
default value is TRUE
It defines that the interpolation method is propagated to interpolated curves.

When set to FALSE the interpolation method is not propagated and the interpolated curve returned does not have interpolation set.

This is necessary for interpolation methods like COPOMScenarios that relies on the original curve.

This slot can be set to FALSE for any interpolation class at any moment.